Adding ProgressBar and SimpleITK events
Utilize SimpleITK event call backs to monitor the filter progress. Add
a qt ProgressBar to report progress to the user.

The python GIL is explicitly yielded to unsure balance between the
main thread and the working thread. So that when a task is added to
the main work queue the main thread has a chance to process before the
returning the filter execution.

The main queue processing is refactored to not be a loop, but to be
placed as an event in the qt application to be called back. This
should balance the work with other qt tasks better. The main thread
explicitly yields the GIL for 0.01 seconds. This allows the SITK
events to acquire the GIL and efficiently run a large number of
event. Since pyQT does not release the GIL, this is required.