Prevent already destroyed widget exception occurring on Slicer exit
This commit avoid the exception reported below from happening when
exiting Slicer:

Exception ValueError: "Trying to call 'deleteLater' on a destroyed qMRMLNodeComboBox object" in <bound method FilterParameters.__del__ of <SimpleFilters.FilterParameters object at 0x5a13390>> ignored

It is not required to explicitly delete the widgets associated with the
last shown filter. The associated widgets have a parent and will be
deleted by the application.